yes 3

+++ src/main/java/com/mystic/atlantis/particles/PushBubbleStreamParticle.java
@@ -1,13 +1,7 @@
 package com.mystic.atlantis.particles;
 
-import com.mojang.math.Vector3f;
-
 import net.minecraft.client.multiplayer.ClientLevel;
-import net.minecraft.client.particle.Particle;
-import net.minecraft.client.particle.ParticleProvider;
-import net.minecraft.client.particle.ParticleRenderType;
-import net.minecraft.client.particle.SpriteSet;
-import net.minecraft.client.particle.TextureSheetParticle;
+import net.minecraft.client.particle.*;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
 import net.minecraft.core.particles.SimpleParticleType;
@@ -15,6 +9,7 @@ import net.minecraft.tags.FluidTags;
 import net.minecraft.util.Mth;
 import net.minecraftforge.api.distmarker.Dist;
 import net.minecraftforge.api.distmarker.OnlyIn;
+import org.joml.Vector3f;
 
 @OnlyIn(Dist.CLIENT)
 public class PushBubbleStreamParticle extends TextureSheetParticle {
@@ -76,7 +71,7 @@ public class PushBubbleStreamParticle extends TextureSheetParticle {
 
          }
          this.move(this.xd, this.yd, this.zd);
-         if (!this.level.getFluidState(new BlockPos(this.x, this.y, this.z)).is(FluidTags.WATER) || this.onGround) {
+         if (!this.level.getFluidState(new BlockPos((int) this.x, (int) this.y, (int) this.z)).is(FluidTags.WATER) || this.onGround) {
             this.remove();
          }
